pyproject.toml
[project]
name = "contextworker"
version = "0.5.0"
description = "The Hands of ContextUnity — Temporal-based durable workflow execution, background jobs, and scheduled task infrastructure"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Pylypchuk Oleksii", email = "oleksii.pylypchuk@tuta.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE.md"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "temporalio>=1.5.0",
    "grpcio>=1.62.0",
    "protobuf>=4.25.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "contextcore>=0.5.1",
]

[project.urls]
Homepage = "https://contextworker.dev"
Documentation = "https://contextworker.dev"
Repository = "https://github.com/ContextUnity/contextworker"
Issues = "https://github.com/ContextUnity/contextworker/issues"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
    "contextcore",
]

